==> 2_find_k.Rmd <==
---
date: "2022-11-08"
output: markdown
---

## Load

```{r load, include=F}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
library(glue)
library(ldatuning)
options(max.print = 1000)
util_folder <- c("util")
beep <- function(){
  system(str_c("notify-send 'Task complete.'",
               "&& paplay /usr/share/sounds/Yaru/stereo/message.oga"))
}
```

```{r read}
data <- read_rds(here("output", "trimmed_data.rds"))
glimpse(data[1:2])
```

```{r corpus}
source(here(util_folder, "count_by_doc.R"))
corpus <- data %>%
  map_dfr(count_by_doc)
glimpse(corpus)
```

```{r dtm}
kxl_dtm <- corpus %>%
    tidytext::cast_dtm(document = id,
                       term = token,
                       value = n)
kxl_dtm
```

## Run function

```{r run}
run_models <- function(data, run, k, seed, overwrite_cache = F){
  if (overwrite_cache){
    results <- ldatuning::FindTopicsNumber(
      kxl_dtm,
      topics = k,
      metrics = c("Griffiths2004", "CaoJuan2009",
                  "Arun2010", "Deveaud2014"),
      method = "Gibbs",
      control = list(seed = seed))
    beep()
    saveRDS(results,
          file = here(".cache", glue("results_{run}.rds")))
  } else {
    results <- read_rds(here(".cache", glue("results_{run}.rds")))
  }
  results
}
```

### Run 1

```{r test_model_1}
results_1 <- run_models(kxl_dtm, "1", c(5, 10, 20, 40, 80, 120), seed = 16869)
ldatuning::FindTopicsNumber_plot(results_1)
```

### Run 2

```{r test_model_2}
results_2 <- run_models(data, "2", seq(10, 50, 10), seed = 222222)
ldatuning::FindTopicsNumber_plot(results_2)
```

### Run 3

```{r test_model_3}
results_3 <- run_models(data, "3", seq(25, 55, 6), seed = 222226, T)
ldatuning::FindTopicsNumber_plot(results_3)
```

### Run 4

```{r test_model_4}
results_4 <- run_models(data, "4", seq(38, 50, 3), seed = 222223)
ldatuning::FindTopicsNumber_plot(results_4)
```

### Run 5

```{r test_model_5}
results_5 <- run_models(data, "5", seq(37, 43), seed = 222224)
ldatuning::FindTopicsNumber_plot(results_5)
```
