==> 3_model.Rmd <==
---
date: "2022-11-08"
output: markdown
---

## Load

```{r load, include=F}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
library(glue)
library(topicmodels)
options(max.print = 1000)
util_folder <- c("util")
beep <- function(){
  system(str_c("notify-send 'Task complete.'",
               "&& paplay /usr/share/sounds/Yaru/stereo/message.oga"))
}
```

```{r read}
data <- read_rds(here("output", "trimmed_data.rds"))
source(here(util_folder, "count_by_doc.R"))
kxl_dtm <- data %>%
  map_dfr(count_by_doc) %>%
  tidytext::cast_dtm(document = id, term = token, value = n)
kxl_dtm
```

```{r data}
glimpse(data[[1]])
```

## Model function

```{r function}
run_model <- function(dtm_input, k_input, seed_input){
  model <- topicmodels::LDA(
    dtm_input,
    k_input,
    method = "Gibbs",
    control = list(seed = seed_input))
  beep()
  saveRDS(model, here(".cache", glue("model_{k_input}.rds")))
  model
}

show_terms <- function(input_model) {
  betas <- tidytext:::tidy.LDA(input_model, matrix = "beta")
  n_topics <- pluck(input_model, "k")
  show_terms_1 <- function(k){
     betas %>%
      filter(topic == k) %>%
      slice_max(beta, n = 20) %>%
      mutate(term = str_trunc(term, 22)) %>%
      print()
  }
  walk(1:n_topics, show_terms_1)
}
```

## Model 1

```{r model_1, eval = F}
model_40 <- run_model(kxl_dtm, 40, 222224)
show_terms(model_40)
```

## Model 2

```{r model_2, eval = F}
model_30 <- run_model(kxl_dtm, 30, 222224)
show_terms(model_30)
```

## Model 3

```{r model_3, eval = F}
model_25 <- run_model(kxl_dtm, 25, 222224)
show_terms(model_25)
```

## Model 4

```{r model_4, eval = F}
model_20 <- run_model(kxl_dtm, 20, 222224)
show_terms(model_20)
```

## Model 4.4

```{r model_4.4, eval=F}
model_17 <- run_model(kxl_dtm, 17, 222224)
show_terms(model_17)
```

## Model 4.5

```{r model_4.5, eval=F}
model_16 <- run_model(kxl_dtm, 16, 222224)
show_terms(model_16)
```

## Model 5

```{r model_5}
model_15 <- run_model(kxl_dtm, 15, 222224)
show_terms(model_15)
```

## Model 6

```{r model_6}
model_14 <- run_model(kxl_dtm, 14, 222224)
show_terms(model_14)
```

## Model 7

```{r model_7}
model_13 <- run_model(kxl_dtm, 13, 222224)
show_terms(model_13)
```

## Model 8

```{r model_8, eval = F}
model_12 <- run_model(kxl_dtm, 12, 222224)
show_terms(model_12)
```

## Model 9

```{r model_9, eval = F}
model_11 <- run_model(kxl_dtm, 11, 222224)
show_terms(model_11)
```

## Final model

```{r gamma}
final_model <- model_14

betas <- tidytext:::tidy.LDA(final_model, matrix = "beta")
glimpse(betas)

get_gammas <- function(tokens, betas){
  tokens %>%
    left_join(betas, by = c("token" = "term")) %>%
    mutate(beta = beta * n) %>%
    group_by(order, person, topic) %>%
    summarize(beta = sum(beta),  .groups = "drop") %>%
    group_by(order, person) %>%
    mutate(total_beta = sum(beta)) %>%
    ungroup() %>%
    mutate(gamma = beta / total_beta) %>%
    select(-c(beta, total_beta, person))
}

complete_entry <- function(entry){
  combinations <- entry$gammas %>%
    full_join(select(entry$document, order), 
              by = c("order")) %>%
    expand(topic, order) %>%
    filter(! is.na(topic)) %>%
    arrange(order)
  entry$gammas <- left_join(combinations, entry$gammas, 
                            by = c("order", "topic")) %>%
    select(order, topic, gamma)
  entry
}

data_betas <- data %>%
  map(~ list_modify(.x, gammas = get_gammas(.x$tokens, betas))) %>%
  map(complete_entry)
glimpse(data_betas[[1]])

write_rds(data_betas, here("output", "data_betas.rds"))
write_rds(final_model, here("output", "topic_model.rds"))
```
